requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
av==12.0.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
Brotli==1.0.9
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
deepspeed==0.17.1
defusedxml==0.7.1
diffusers==0.33.0
docopt==0.6.2
einops==0.8.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
fonttools==4.58.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.2
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
gmpy2==2.2.1
h11==0.16.0
hf-xet==1.1.3
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.7
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
lightning-utilities==0.14.3
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mistune==3.1.3
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.4.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.4
notebook==7.4.1
notebook_shim==0.2.4
numpy==2.0.1
nvidia-ml-py==12.575.51
omegaconf==2.3.0
opencv-python==4.11.0.86
overrides==7.7.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
peft==0.15.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==25.1
pipdeptree==2.26.1
pipreqs==0.5.0
platformdirs==4.3.7
proglog==0.1.11
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pytorch-lightning==2.5.1
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
safetensors==0.5.3
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.30.0
setproctitle==1.3.6
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.1
terminado==0.18.1
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
torchmetrics==1.7.3
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.1
triton==3.1.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
typing-inspection==0.4.1
uri-template==1.3.0
urllib3==2.3.0
wandb==0.20.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
yarg==0.1.9
yarl==1.20.1
zipp==3.21.0
